Address CodeRabbit findings: test name, duplicate test, redundant cast
- Renamed test_claim_verifier_records_force_stop_verdict_source_for_plain_text_recovery
  to test_claim_verifier_recovers_from_plain_text_with_json_verdict (the
  test asserts verdict_source=="json_schema", not "force_stop").
- Removed test_extract_json_verdict_returns_none_without_parsed_json —
  exact duplicate of the first assertion in
  test_extract_json_verdict_requires_valid_parsed_json.
- Removed redundant int() wrapping around round() in
  _resolve_coordinates (round() with one arg returns int in Python 3).
- mouse_down/mouse_up overlay preview: acknowledged as cosmetic; the
  preview call is already made but the overlay silently ignores
  unknown action types. Adding handlers is visual polish, not a
  correctness issue.

225 tests pass.

+++ src/frontend_visualqa/actions.py
@@ -574,7 +574,7 @@ class ActionExecutor:
             if result.get("success"):
                 resolved_coordinates = result.get("coordinates")
                 if isinstance(resolved_coordinates, (list, tuple)) and len(resolved_coordinates) == 2:
-                    return int(round(float(resolved_coordinates[0]))), int(round(float(resolved_coordinates[1])))
+                    return round(float(resolved_coordinates[0])), round(float(resolved_coordinates[1]))
             if not has_coordinates:
                 message = result.get("message", "Unknown error")
                 raise BrowserActionError(f"{action_name} ref resolution failed for {ref}: {message}")
